+++ jagua-rs/src/probs/spp/io/import.rs
@@ -38,8 +38,8 @@ pub fn import(importer: &Importer, ext_instance: &ExtSPInstance) -> Result<SPIns
 
     let total_item_area = items
         .iter()
-        .map(|(item, demand)| item.area() * *demand as f32)
-        .sum::<f32>();
+        .map(|(item, demand)| item.area() * *demand as f64)
+        .sum::<f64>();
 
     let fixed_height = ext_instance.strip_height;
